+++ src/components/CallNotify.tsx
@@ -74,11 +74,12 @@ export const CallNotify = observer(() => {
     !getAuthStore().getCurrentAccount()?.pushNotificationEnabled
       ? Fragment
       : DidMountTimer
-
   const configure = getAuthStore().pbxConfig
   const hideHangup =
     c.incoming && configure?.['webphone.call.hangup'] === 'false'
-  const numberCalls = getCallStore().calls?.length || 0
+  const n = getCallStore().calls.filter(
+    _ => _.incoming && !_.answered && _.id !== c.id,
+  ).length
   return (
     <Wrapper>
       {getCallStore().shouldRingInNotify(c.callkeepUuid) && <IncomingItem />}
@@ -90,7 +91,7 @@ export const CallNotify = observer(() => {
           <RnText bold>{c.getDisplayName()}</RnText>
           <RnText>
             {intl`Incoming Call`}
-            {numberCalls > 1 ? ` (${numberCalls})` : ''}
+            {n > 0 ? ' (' + intl`${n} in background` + ')' : ''}
           </RnText>
         </View>
         {!hideHangup && (